# sync(probspecs): refactor