--- a/src/sync/probspecs.nim
+++ b/src/sync/probspecs.nim
@@ -29,11 +29,8 @@ proc clone(probSpecsDir: ProbSpecsDir) =
   logNormal(&"Cloning the problem-specifications repo into {probSpecsDir}...")
   execCmdException(cmd, "Could not clone problem-specifications repo")
 
-func exercisesDir(probSpecsDir: ProbSpecsDir): string =
-  probSpecsDir / "exercises"
-
 proc exercises(probSpecsDir: ProbSpecsDir): seq[ProbSpecsExerciseDir] =
-  for exerciseDir in walkDirs(probSpecsDir.exercisesDir / "*"):
+  for exerciseDir in walkDirs(probSpecsDir / "exercises" / "*"):
     result.add ProbSpecsExerciseDir(exerciseDir)
 
 func canonicalDataFile(probSpecsExerciseDir: ProbSpecsExerciseDir): string =