This PR is part of a series that addresses #64.

I've verified that this commit keeps the output of
`configlet sync --check` the same on every track.

This PR is currently a 3–⁠5x speedup, depending on the track.

Measuring the execution time of
```
./configlet sync --check --offline --probSpecsDir=/path/to/local/problem-specifications
```

in different track repos, before and after this PR:

|        |    ada |    nim |   csharp |
| ------ | -----: | -----: | -------: |
| Before | 100 ms | 110 ms |   120 ms |
| After  |  20 ms |  35 ms |    40 ms |

It wasn't particularly slow before for the end-user, but it was
noticeable when running `configlet sync --check` on every track for
regression testing).

An upcoming commit will optimize further: currently we parse every
`canonical-data.json` file in `problem-specifications` even when the
track has no Practice Exercises (like the `ada` track).

With this commit, most of the runtime of `configlet sync --check` is now
spent parsing TOML and JSON, as you'd expect.

See the profiling below for the Nim track, where there are 67 practice
exercises that have a corresponding `canonical-data.json` in
`problem-specifications`. We can see the 67 calls to `parsetoml`, but
119 calls to `parseJson` (and also the special handling of the `grains`
exercise).

![prof1](https://user-images.githubusercontent.com/45465154/119169911-d818d700-ba62-11eb-9942-173c033858d8.png)